mv ReadGPLink logic to helper file
