+++ src/CommonLib/Processors/ContainerProcessor.cs
@@ -121,11 +121,7 @@ namespace SharpHoundCommonLib.Processors
 
         public IAsyncEnumerable<GPLink> ReadContainerGPLinks(ResolvedSearchResult result, IDirectoryObject entry)
         {
-            if (entry.TryGetProperty(LDAPProperties.GPLink, out var links)) {
-                return ReadContainerGPLinks(links);    
-            }
-
-            return AsyncEnumerable.Empty<GPLink>();
+            return Helpers.ReadGPLinks(entry, _utils);
         }
 
         /// <summary>
@@ -133,25 +129,9 @@ namespace SharpHoundCommonLib.Processors
         /// </summary>
         /// <param name="gpLink"></param>
         /// <returns></returns>
-        public async IAsyncEnumerable<GPLink> ReadContainerGPLinks(string gpLink)
+        public IAsyncEnumerable<GPLink> ReadContainerGPLinks(string gpLink)
         {
-            if (gpLink == null)
-                yield break;
-
-            foreach (var link in Helpers.SplitGPLinkProperty(gpLink))
-            {
-                var enforced = link.Status.Equals("2");
-
-                var res = await _utils.ResolveDistinguishedName(link.DistinguishedName);
-
-                if (res.Success) {
-                    yield return new GPLink
-                    {
-                        GUID = res.Principal.ObjectIdentifier,
-                        IsEnforced = enforced
-                    };
-                }
-            }
+            return Helpers.ReadGPLinks(gpLink, _utils);
         }
 
         /// <summary>
